Move `VariantValidationResult` to models, and freeze it
